[apps/infra] resourceVersion implementation in progress
