+++ apps/infra/internal/domain/domain.go
@@ -2,8 +2,11 @@ package domain
 
 import (
 	"encoding/json"
+	"fmt"
+	"strconv"
 
 	"github.com/kloudlite/operator/agent/types"
+	"github.com/kloudlite/operator/pkg/constants"
 	"github.com/kloudlite/operator/pkg/kubectl"
 	"go.uber.org/fx"
 	"kloudlite.io/apps/infra/internal/domain/entities"
@@ -62,7 +65,16 @@ func (d *domain) dispatchToTargetAgent(ctx InfraContext, action agent.Action, cl
 	return err
 }
 
-func (d *domain) applyK8sResource(ctx InfraContext, obj client.Object) error {
+func (d *domain) applyK8sResource(ctx InfraContext, obj client.Object, recordVersion int) error {
+	if recordVersion > 0 {
+		ann := obj.GetAnnotations()
+		if ann == nil {
+			ann = make(map[string]string, 1)
+		}
+		ann[constants.RecordVersionKey] = fmt.Sprintf("%d", recordVersion)
+		obj.SetAnnotations(ann)
+	}
+
 	b, err := fn.K8sObjToYAML(obj)
 	if err != nil {
 		return err
@@ -84,6 +96,20 @@ func (d *domain) deleteK8sResource(ctx InfraContext, obj client.Object) error {
 		return err
 	}
 	return nil
+}
+
+func (d *domain) parseRecordVersionFromAnnotations(annotations map[string]string) (int, error) {
+	annotatedVersion, ok := annotations[constants.RecordVersionKey]
+	if !ok {
+		return 0, fmt.Errorf("no annotation with record version key (%s), found on the resource", constants.RecordVersionKey)
+	}
+
+	annVersion, err := strconv.ParseInt(annotatedVersion, 10, 32)
+	if err != nil {
+		return 0, err
+	}
+
+	return int(annVersion), nil
 }
 
 var Module = fx.Module("domain",
